# public-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kitchen Master API",
        "title": "API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/clientuser": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get client user info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClientUser"
                ],
                "summary": "Get client user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ClientUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update client user info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClientUser"
                ],
                "summary": "Update client user info",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.clientUserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete all client user data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClientUser"
                ],
                "summary": "Delete all client user data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "required": [
                "content",
                "contentType",
                "status"
            ],
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "contentType": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "api.clientUserUpdateRequest": {
            "type": "object",
            "required": [
                "allergies",
                "autoImportEnabled",
                "calories",
                "carbohydrates",
                "fats",
                "favoriteRecipeIDs",
                "height",
                "protein",
                "username",
                "weight"
            ],
            "properties": {
                "allergies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "autoImportEnabled": {
                    "type": "boolean"
                },
                "calories": {
                    "type": "number"
                },
                "carbohydrates": {
                    "type": "number"
                },
                "fats": {
                    "type": "number"
                },
                "favoriteRecipeIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "protein": {
                    "type": "number"
                },
                "username": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
        "entity.ClientUser": {
            "type": "object",
            "required": [
                "Id",
                "allergies",
                "autoImportEnabled",
                "billing",
                "calories",
                "carbohydrates",
                "elasticIndex",
                "email",
                "fats",
                "favoriteRecipeIDs",
                "height",
                "importedRecipes",
                "protein",
                "subscription",
                "username",
                "weight"
            ],
            "properties": {
                "Id": {
                    "type": "string"
                },
                "allergies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "autoImportEnabled": {
                    "description": "NOTE: if auto imported is enabled we import new recipe files automatically",
                    "type": "boolean"
                },
                "billing": {
                    "type": "number"
                },
                "calories": {
                    "type": "number"
                },
                "carbohydrates": {
                    "type": "number"
                },
                "elasticIndex": {
                    "description": "NOTE: index name in elastic search",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fats": {
                    "type": "number"
                },
                "favoriteRecipeIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "importedRecipes": {
                    "description": "NOTE: keeps track of imported files",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "protein": {
                    "type": "number"
                },
                "subscription": {
                    "description": "NOTE: FREE subscription has limited access",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}